+++ pipeline/train_instruct.py
@@ -15,7 +15,6 @@ Launch:
 
 import json
 import os
-import re
 import sys
 import uuid
 from contextlib import nullcontext
@@ -47,62 +46,17 @@ from config.training_config import InstructConfig
 from models.tiny_aya_vision import TinyAyaVisionForConditionalGeneration
 from pipeline.data import InstructDataset, collate_fn
 from pipeline.apply_lora import apply_lora, get_lora_optimizer_groups
+from pipeline.utils import (
+    is_torchrun,
+    setup_ddp,
+    cleanup_ddp,
+    _unwrap_model,
+    save_checkpoint,
+    find_latest_checkpoint,
+    save_hf_model,
+    build_lr_scheduler,
+)
 from src.processing import TinyAyaVisionProcessor
-from models import save_for_inference
-
-
-def is_torchrun() -> bool:
-    """True when launched via torchrun / torch.distributed.launch."""
-    return "LOCAL_RANK" in os.environ
-
-
-def setup_ddp():
-    """Initialize distributed process group and set CUDA device."""
-    local_rank = int(os.environ["LOCAL_RANK"])
-    torch.cuda.set_device(local_rank)
-    dist.init_process_group(backend="nccl", device_id=torch.device(f"cuda:{local_rank}"))
-    return local_rank
-
-
-def cleanup_ddp():
-    if dist.is_initialized():
-        dist.destroy_process_group()
-
-
-def _unwrap_model(model):
-    """Unwrap torch.compile and DDP wrappers to access the raw module."""
-    raw = model
-    if hasattr(raw, "_orig_mod"):    # torch.compile
-        raw = raw._orig_mod
-    if hasattr(raw, "module"):       # DDP
-        raw = raw.module
-    return raw
-
-
-def save_checkpoint(checkpoint_dir, step, model, optimizer, lr_scheduler):
-    save_path = checkpoint_dir / f"checkpoint_{step}.pt"
-    raw_model = _unwrap_model(model)
-    torch.save({
-        "step": step,
-        "projector": raw_model.multi_modal_projector.state_dict(),
-        "lora_adapter": {
-            k: v for k, v in raw_model.language_model.state_dict().items()
-            if "lora_" in k
-        },
-        "optimizer": optimizer.state_dict(),
-        "lr_scheduler": lr_scheduler.state_dict(),
-    }, save_path)
-    print(f"Saved checkpoint to {save_path}")
-
-
-def find_latest_checkpoint(checkpoint_dir: Path) -> Path | None:
-    checkpoints = list(checkpoint_dir.glob("checkpoint_*.pt"))
-    if not checkpoints:
-        return None
-    def extract_step(p):
-        m = re.search(r"checkpoint_(\d+)\.pt$", p.name)
-        return int(m.group(1)) if m else -1
-    return max(checkpoints, key=extract_step)
 
 
 @torch.no_grad()
@@ -346,7 +300,7 @@ def train(
                         tqdm.write(f"Epoch {epoch}, Opt Step {opt_step}, Loss {_loss_val:.4f}, LR {_lr_val}")
 
                     if _should_save:
-                        save_checkpoint(checkpoint_dir, step + 1, model, optimizer, lr_scheduler)
+                        save_checkpoint(checkpoint_dir, step + 1, model, optimizer, lr_scheduler, save_lora=True)
 
                         if processor is not None:
                             samples = generate_samples(
@@ -368,7 +322,7 @@ def train(
 
     hook_handle.remove()
     if is_main:
-        save_checkpoint(checkpoint_dir, step + 1, model, optimizer, lr_scheduler)
+        save_checkpoint(checkpoint_dir, step + 1, model, optimizer, lr_scheduler, save_lora=True)
     if use_ddp:
         dist.barrier()
     if is_main:
@@ -551,22 +505,7 @@ def main(
         weight_decay=training_config.weight_decay,
     )
 
-    full_loader_len = full_dataset_len // (per_gpu_batch_size * world_size)
-    total_steps = training_config.num_epochs * full_loader_len // training_config.grad_acc_steps
-    warmup_steps = int(total_steps * training_config.warmup_ratio)
-
-    if training_config.lr_scheduler_type == "cosine":
-        warmup_scheduler = torch.optim.lr_scheduler.LinearLR(
-            opt, start_factor=1e-8 / training_config.learning_rate, total_iters=warmup_steps,
-        )
-        cosine_scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(
-            opt, T_max=total_steps - warmup_steps, eta_min=training_config.learning_rate * 0.01,
-        )
-        lr_scheduler = torch.optim.lr_scheduler.SequentialLR(
-            opt, schedulers=[warmup_scheduler, cosine_scheduler], milestones=[warmup_steps],
-        )
-    else:
-        raise ValueError(f"Unsupported LR scheduler type: {training_config.lr_scheduler_type}")
+    lr_scheduler = build_lr_scheduler(opt, training_config, full_dataset_len, per_gpu_batch_size, world_size)
 
     if resume_step > 0:
         opt.load_state_dict(ckpt["optimizer"])
@@ -588,13 +527,7 @@ def main(
     )
 
     if is_main and training_config.save_hf_model:
-        print("Merging LoRA and saving HF-compatible model...")
-        raw_model = _unwrap_model(model)
-        raw_model.language_model = raw_model.language_model.merge_and_unload()
-
-        hf_output_dir = checkpoint_dir / "hf_model"
-        save_for_inference(raw_model, processor, hf_output_dir)
-        print(f"Saved HF-compatible model to {hf_output_dir}")
+        save_hf_model(model, processor, checkpoint_dir)
 
     if is_main:
         wandb.finish()
